Create entropy_avail
This algorithm was used to obtain the entropy available in the kernel every second. This code was written on C.
This was used to inspect how the entropy available at the moment was altering due to any hardware or kernel interrupts.
Entropy available stored in a file was accessed through read mode.
Every 1 second it was displayed on the terminal 
It showed only 256 bits every second as the real value is being masked by the software for security reasons.