Create Kernel Data Structures
This kernel module is designed to get and store fundamental information about the currently executing process at the kernel level. It connects to the Linux kernel with the help of keys like __innit and cleanup __exit functions, which are automatically called during module insertion and removal.
Upon initialization, the module uses the task_struct of the current process via the “current” macro. This process retrieves PID, UID, Process Name and Total virtual memory.
It uses kernel APIs such as get_task_cred() and memory descriptor checks mm_struct() to safely fetch and display the relevant process and memory details. All the output is logged using printk() with the KERN_INFO log level, hence, the data can be displayed through dmesg or kernel logs for debugging purposes